requirements.txt
streamlit==1.38.0
streamlit-extras==0.4.7
reportlab==4.2.2
pillow==10.4.0
python-dotenv==1.0.1
streamlit==1.39.0
streamlit-extras==0.4.7
reportlab==4.2.2
pillow==10.4.0
pytz==2024.2
numpy==2.0.2
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2